requirements.txt
BeautifulSoup
chardet
lxml
readability-lxml
requests
wsgiref
